% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_fires.R
\name{ef_fires}
\alias{ef_fires}
\title{Search the European Forest Fire Information System 'EFFIS'}
\usage{
ef_fires(country_iso2 = NULL, limit = NULL, updated = NULL,
  detected = NULL, fireId = NULL, page = NULL, ...)
}
\arguments{
\item{country_iso2}{(character) ISO2 country code}

\item{limit}{(integer) record limit}

\item{updated}{(character) date time}

\item{detected}{(character) date time}

\item{fireId}{(character) Fire ID}

\item{page}{(integer) Page number}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search the European Forest Fire Information System 'EFFIS'
}
