% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_hcu.R
\name{ef_hcu}
\alias{ef_hcu}
\title{Search the European Forest Fire Information System 'EFFIS'}
\usage{
ef_hcu(country, year)
}
\arguments{
\item{country}{(country) ISO2 code, only one country}

\item{year}{(integer) One or more year, must be smaler than the current year}
}
\description{
Search the European Forest Fire Information System 'EFFIS'
}
