% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_trend.R
\name{ef_trend}
\alias{ef_trend}
\title{Search the European Forest Fire Information System 'EFFIS'}
\usage{
ef_trend(country = NULL, province = NULL, limit = NULL, firedate = NULL,
  area_ha = NULL, ba_class = NULL, decimate = NULL, page = NULL, ...)
}
\arguments{
\item{country}{(character) ISO country code}

\item{province}{(character) Province name}

\item{limit}{(integer) record limit}

\item{firedate}{(character) YYYY-MM-DD}

\item{area_ha}{(integer) size}

\item{ba_class}{(character) one of 'ALL', '30DAYS' or '07DAYS'}

\item{decimate}{(integer) pick one date every x}

\item{page}{(integer) Page number}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search the European Forest Fire Information System 'EFFIS'
}
