% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effisr-search.R
\name{eff_search}
\alias{eff_search}
\title{Search EFFIS}
\usage{
eff_search(country = NULL, countryful = NULL, province = NULL,
  commune = NULL, limit = NULL, critech = NULL, firedate = NULL,
  area_ha = NULL, page = NULL, ...)
}
\arguments{
\item{country}{(character) Country name, a 2 letter code}

\item{countryful}{(character) Country name, long name}

\item{province}{(character) Province name}

\item{commune}{(character) Commune}

\item{limit}{(integer) record limit}

\item{critech}{(character) one of 'YES'}

\item{firedate}{(character) YYYy-MM-DD}

\item{area_ha}{(integer) size}

\item{page}{(integer) Page number}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search EFFIS
}
