% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_hcu.R
\name{ef_hcu}
\alias{ef_hcu}
\title{Search the European Forest Fire Information System 'EFFIS'}
\usage{
ef_hcu(year, country)
}
\arguments{
\item{year}{(integer) One or more year, must be smaler than the current year}

\item{country}{(country) ISO2 code, only one country}
}
\description{
Search the European Forest Fire Information System 'EFFIS'
}
