% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ef_trend.R
\name{ef_trend}
\alias{ef_trend}
\title{Search the European Forest Fire Information System 'EFFIS'}
\usage{
ef_trend(country = NULL, province = NULL, limit = NULL,
  firedate = NULL, area_ha = NULL, ba_class = NULL,
  ordering = NULL, decimate = NULL, page = NULL, ...)
}
\arguments{
\item{country}{(character) One or more country names as ISO2 code}

\item{province}{(character) Province name}

\item{limit}{(integer) Record limit}

\item{firedate}{(character) YYYY-MM-DD}

\item{area_ha}{(integer) Size of the fire}

\item{ba_class}{(character) One of 'ALL', '30DAYS' or '07DAYS'}

\item{ordering}{(character) One or more field names. Specifies the sort order.
The names can be optionally prefixed by “-” to indicate descending sort.}

\item{decimate}{(integer) pick one date every x}

\item{page}{(integer) Page number}

\item{...}{Further args passed on to \code{\link[httr]{GET}}}
}
\description{
Search the European Forest Fire Information System 'EFFIS'
}
