% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effisr-search.R
\name{eff_search}
\alias{eff_search}
\title{Search the European Forest Fire Information System 'EFFIS'}
\usage{
eff_search(country = NULL, province = NULL, limit = NULL,
  firedate = NULL, area_ha = NULL, ba_class = NULL, page = NULL, ...)
}
\arguments{
\item{country}{(character) ISO country code}

\item{province}{(character) Province name}

\item{limit}{(integer) record limit}

\item{firedate}{(character) YYYY-MM-DD}

\item{area_ha}{(integer) size}

\item{ba_class}{(character) one of 'ALL', '30DAYS' or '07DAYS'}

\item{page}{(integer) Page number}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\description{
Search the European Forest Fire Information System 'EFFIS'
}
