% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effisr-package.R
\docType{package}
\name{effisr-package}
\alias{effisr-package}
\alias{effisr}
\title{R client for EFFIS}
\description{
R Client for the European Forest Fire Information System (EFFIS) API
}
\section{About}{


This package gives you access to data from EFFIS System
\url{http://effis.jrc.ec.europa.eu/} via their API
(\url{http://effis.jrc.ec.europa.eu/api-test/})
}

\section{Functions}{


\itemize{
 \item \code{\link{ef_current}}
 \item \code{\link{ef_trend}}
 \item \code{\link{ef_fires}}
 \item \code{\link{ef_hcu}}
}


- URL: <https://github.com/patperu/effisr>
- BugReports: <https://github.com/patperu/effisr/issues>
}

\author{
Patrick Hausmann \email{patrick.hausmann@covimo.de}
}
\keyword{package}
