% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effis-package.R
\docType{package}
\name{effisr-package}
\alias{effisr-package}
\alias{effisr}
\title{R client for Effis}
\description{
R client for Effis
}
\author{
Patrick Hausmann \email{patrick.hausmann@covimo.de}
}
\keyword{package}
